refactor(scripts): split bundler into library + CLI; portable isMain() in upload wrapper
- The import.meta.main guard used by both scripts returned undefined under
  tsx (v4.21.0, Node 22), so `pnpm tsx scripts/*.ts` never executed the
  CLI path. Replace with a portable isMain() that compares import.meta.url
  to pathToFileURL(process.argv[1]).href and falls back gracefully.
- Split scripts/bundle-hubspot-card.ts into a library (bundleTargets,
  buildViteOptions — pure, no plugin imports) and a new CLI file
  scripts/bundle-hubspot-card-cli.ts (static import of @vitejs/plugin-react
  + vite). This makes the library importable from tests without requiring
  @vitejs/plugin-react to be resolvable from the repo root.
- scripts/hs-project-upload.ts now invokes bundle-hubspot-card-cli.ts.
- Harden the noNonNullAssertion lint rule in the new bundler tests.
